Move files and add initial redux config
