+++ 2.Front-end-Development/exercises/Unit_16/Day_3/class-exercises/my-forms-2.0/src/components/state.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import dataStates from './dataStates'
+import dataStates from '../data/dataStates'
 
 class State extends React.Component {
   render() {
