+++ 2.Front-end-Development/exercises/Unit_16/Day_3/class-exercises/my-forms-2.0/src/redux/reducers/formReducer.js
@@ -1,0 +1,20 @@
+const INITIAL_STATE = {
+  name: '',
+  email: '',
+  cpf: '',
+  address: '',
+  city: '',
+  state: '',
+  homeType: '',
+  professionalResume: '',
+  occupation: '',
+  wasAlerted: false,
+  jobDescription: '',
+  isSend: false,
+}
+
+const formReducer = (state = INITIAL_STATE, action) => {
+  return state
+}
+
+export default formReducer;
