+++ 2.Front-end-Development/exercises/Unit_16/Day_3/class-exercises/my-forms-2.0/src/redux/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import formReducer from './formReducer';
+
+const rootReducer = combineReducers({ formReducer });
+
+export default rootReducer;
