+++ 2.Front-end-Development/exercises/Unit_16/Day_3/class-exercises/my-forms-2.0/src/redux/store/index.js
@@ -1,0 +1,9 @@
+import { createStore } from 'redux';
+import rootReducer from '../reducers';
+
+const store = createStore(
+  rootReducer,
+  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+);
+
+export default store;
